back-end/requirements.txt
# Existing dependencies (from your current setup)
fastapi
uvicorn[standard]
pydantic
python-multipart
aiofiles
aiohttp
python-dotenv
beautifulsoup4
requests

# Security and content sanitization
bleach>=6.1.0

# New dependencies for Bayesian model
numpy>=1.21.0
scipy>=1.7.0
schedule>=1.2.0

# Optional: for enhanced mathematical operations
pandas>=1.3.0  # if you want to use pandas for data processing
matplotlib>=3.5.0  # if you want to generate charts
seaborn>=0.11.0  # for enhanced visualization (optional)

# ETH Kalman filter dependencies
filterpy>=1.4.5
websockets>=12.0

# XAU (Gold) price data (API Ninjas)
# Uses aiohttp which is already listed above

# Auth & security
argon2-cffi>=23.1.0
PyJWT>=2.8.0

# Health module dependencies
fastdtw>=0.3.4

# OpenAI for ChatGPT news summarization
openai>=1.0.0

# Testing dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.27.0  # FastAPI test client
faker>=22.0.0  # Generate test data
pytest-mock>=3.12.0  # Mocking utilities
pytest-timeout>=2.2.0  # Test timeout handling
